수업 완

+++ src/main/java/com/green/greengram/application/feed/FeedRepository.java
@@ -3,6 +3,6 @@ package com.green.greengram.application.feed;
 import com.green.greengram.entity.Feed;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedRepository extends JpaRepository<Feed, Integer> {
+public interface FeedRepository extends JpaRepository<Feed, Long> {
 
 }
